resolved build fail issues
